meditrcak/pyproject.toml
[project]
name = "meditrcak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==3.2.2",
    "elevenlabs>=2.27.0",
    "email-validator==2.3.0",
    "faiss-cpu>=1.13.2",
    "fastapi==0.123.8",
    "groq>=0.37.1",
    "gtts>=2.5.4",
    "httpx==0.28.1",
    "jinja2==3.1.6",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langgraph>=1.0.5",
    "numpy>=2.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pathlib2==2.3.7.post1",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "pypdf>=6.5.0",
    "pytest==9.0.1",
    "pytest-asyncio==1.3.0",
    "python-dotenv==1.1.1",
    "python-jose[cryptography]==3.5.0",
    "python-multipart==0.0.20",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "scipy>=1.16.3",
    "sentence-transformers>=5.2.0",
    "sqlalchemy==2.0.41",
    "streamlit>=1.52.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "twilio==9.8.8",
    "uvicorn==0.38.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
